+++ multiqc/modules/picard/MarkDuplicates.py
@@ -28,16 +28,10 @@ def parse_reports(self):
                 s_name = None
 
                 # Pull sample name from input
-                fn_search = re.search(r"INPUT=(\[?[^\s]+\]?)", l, flags=re.IGNORECASE)
+                fn_search = re.search(r"INPUT=?\s*(\[?[^\s]+\]?)", l, flags=re.IGNORECASE)
                 if fn_search:
                     s_name = os.path.basename(fn_search.group(1).strip('[]'))
                     s_name = self.clean_s_name(s_name, f['root'])
-                # When run with GATK this has a different format
-                else:
-                    fn_search = re.search(r"--input (\[?[^\s]+\]?)", l, flags=re.IGNORECASE)
-                    if fn_search:
-                        s_name = os.path.basename(fn_search.group(1).strip('[]'))
-                        s_name = self.clean_s_name(s_name, f['root'])
 
             if s_name is not None:
                 if 'DuplicationMetrics' in l and '## METRICS CLASS' in l: