+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 * **Peddy**
     * Switch `Sex error` logic to `Correct sex` for better highlighting ([@aledj2](https://github.com/aledj2))
 * **Picard**
-    * Updated module and search patterns to recognise new output format from Picard version >= 2.16
+    * Updated module and search patterns to recognise new output format from Picard version >= 2.16 and GATK output.
 * **RNA-SeQC**
     * Added rRNA alignment stats to summary table [@Rolandde](https://github.com/Rolandde)
 * **RSeqC**